fixing jenkins warning.

+++ opm/autodiff/MSWellHelpers.hpp
@@ -50,8 +50,8 @@ namespace mswellhelpers
 
         // Checking if there is any inf or nan in y
         // it will be the solution before we find a way to catch the singularity of the matrix
-        for (int i_block = 0; i_block < y.size(); ++i_block) {
-            for (int i_elem = 0; i_elem < y[i_block].size(); ++i_elem) {
+        for (size_t i_block = 0; i_block < y.size(); ++i_block) {
+            for (size_t i_elem = 0; i_elem < y[i_block].size(); ++i_elem) {
                 if (std::isinf(y[i_block][i_elem]) || std::isnan(y[i_block][i_elem]) ) {
                     OPM_THROW(Opm::NumericalProblem, "nan or inf value found in invDXDirect due to singular matrix");
                 }
